refactor: simplify method and file names
