+++ src/__tests__/feature-collection.test.ts
@@ -1,12 +1,13 @@
 import { Feature, Point } from "geojson"
 
-import { transformRecordsToFeatureCollection } from "../transform-records"
+import { createFeatureCollection } from "../feature-collection"
 import { fixture as validRecordsFixture } from "../__fixtures__/valid-records"
 import { fixture as customColumnsFixture } from "../__fixtures__/custom-column"
 import { fixture as invalidGeocodeFixture } from "../__fixtures__/invalid-geocode"
 import { fixture as missingGeocodeFixture } from "../__fixtures__/missing-geocode"
+import { StringMap } from "../types"
 
-interface FixtureFields {
+interface FixtureFields extends StringMap {
   Name: string
   "Zip Code": string
 }
@@ -15,13 +16,13 @@ const deepRecreate = (obj: unknown) => JSON.parse(JSON.stringify(obj))
 
 let records: Airtable.Records<FixtureFields>
 
-describe(transformRecordsToFeatureCollection, () => {
+describe(createFeatureCollection, () => {
   beforeEach(() => {
     records = deepRecreate(validRecordsFixture)
   })
 
   it("returns a FeatureCollection object and an errors object", () => {
-    const result = transformRecordsToFeatureCollection(records)
+    const result = createFeatureCollection(records)
 
     expect(result).toBeInstanceOf(Array)
 
@@ -43,9 +44,7 @@ describe(transformRecordsToFeatureCollection, () => {
   })
 
   it("transforms the Airtable records & fields into GeoJSON features & properties", () => {
-    const [featureCollection, _errors] = transformRecordsToFeatureCollection(
-      records
-    )
+    const [featureCollection, _errors] = createFeatureCollection(records)
 
     expect(featureCollection.features).toHaveLength(3)
 
@@ -62,9 +61,7 @@ describe(transformRecordsToFeatureCollection, () => {
   })
 
   it("decodes the Airtable cached geocode field into a GeoJSON geometry", () => {
-    const [featureCollection, _errors] = transformRecordsToFeatureCollection(
-      records
-    )
+    const [featureCollection, _errors] = createFeatureCollection(records)
 
     const feature0 = featureCollection.features[0] as Feature<Point>
     expect(feature0.geometry.coordinates[0]).toBeCloseTo(-73.932376, 6)
@@ -80,9 +77,7 @@ describe(transformRecordsToFeatureCollection, () => {
   })
 
   it("removes the encoded geodata field", () => {
-    const [featureCollection, _errors] = transformRecordsToFeatureCollection(
-      records
-    )
+    const [featureCollection, _errors] = createFeatureCollection(records)
 
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const f = featureCollection.features[0] as Feature<Point, any>
@@ -92,9 +87,7 @@ describe(transformRecordsToFeatureCollection, () => {
   })
 
   it("treats the record id as the feature id", () => {
-    const [featureCollection, _errors] = transformRecordsToFeatureCollection(
-      records
-    )
+    const [featureCollection, _errors] = createFeatureCollection(records)
 
     records.map((record, i) => {
       expect(featureCollection.features[i].id).toEqual(record.id)
@@ -102,7 +95,7 @@ describe(transformRecordsToFeatureCollection, () => {
   })
 
   it("renders the correct json", () => {
-    expect(transformRecordsToFeatureCollection(records)).toMatchInlineSnapshot(`
+    expect(createFeatureCollection(records)).toMatchInlineSnapshot(`
       Array [
         Object {
           "features": Array [
@@ -168,12 +161,9 @@ describe(transformRecordsToFeatureCollection, () => {
     })
 
     it("decodes the geodata", () => {
-      const [featureCollection, _errors] = transformRecordsToFeatureCollection(
-        records,
-        {
-          geocodedFieldName: "Geocodez",
-        }
-      )
+      const [featureCollection, _errors] = createFeatureCollection(records, {
+        geocodedFieldName: "Geocodez",
+      })
 
       const feature0 = featureCollection.features[0] as Feature<Point>
       expect(feature0.geometry.coordinates[0]).toBeCloseTo(-73.932376, 6)
@@ -181,12 +171,9 @@ describe(transformRecordsToFeatureCollection, () => {
     })
 
     it("removes the encoded geodata field", () => {
-      const [featureCollection, _errors] = transformRecordsToFeatureCollection(
-        records,
-        {
-          geocodedFieldName: "Geocodez",
-        }
-      )
+      const [featureCollection, _errors] = createFeatureCollection(records, {
+        geocodedFieldName: "Geocodez",
+      })
 
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       const f = featureCollection.features[0] as Feature<Point, any>
@@ -209,7 +196,7 @@ describe(transformRecordsToFeatureCollection, () => {
     })
 
     it("logs a message to the console", () => {
-      transformRecordsToFeatureCollection(records)
+      createFeatureCollection(records)
       expect(console.error).toHaveBeenCalled()
       expect((console.error as jest.Mock).mock.calls[0][0]).toMatch(
         /LatLng appears to be undefined/
@@ -217,16 +204,12 @@ describe(transformRecordsToFeatureCollection, () => {
     })
 
     it("returns the bad record in the errors object", () => {
-      const [_featureCollection, errors] = transformRecordsToFeatureCollection(
-        records
-      )
+      const [_featureCollection, errors] = createFeatureCollection(records)
       expect(errors.invalidGeocodes).toEqual([records[0]])
     })
 
     it("returns other valid features", () => {
-      const [featureCollection, _errors] = transformRecordsToFeatureCollection(
-        records
-      )
+      const [featureCollection, _errors] = createFeatureCollection(records)
       expect(records).toHaveLength(2)
       expect(featureCollection.features).toHaveLength(1)
     })
@@ -245,7 +228,7 @@ describe(transformRecordsToFeatureCollection, () => {
     })
 
     it("logs a message to the console", () => {
-      transformRecordsToFeatureCollection(records)
+      createFeatureCollection(records)
       expect(console.error).toHaveBeenCalled()
       expect((console.error as jest.Mock).mock.calls[0][0]).toMatch(
         /geocoded field value was missing/
@@ -253,16 +236,12 @@ describe(transformRecordsToFeatureCollection, () => {
     })
 
     it("returns the bad record in the errors object", () => {
-      const [_featureCollection, errors] = transformRecordsToFeatureCollection(
-        records
-      )
+      const [_featureCollection, errors] = createFeatureCollection(records)
       expect(errors.missingGeocodes).toEqual([records[0]])
     })
 
     it("returns other valid features", () => {
-      const [featureCollection, _errors] = transformRecordsToFeatureCollection(
-        records
-      )
+      const [featureCollection, _errors] = createFeatureCollection(records)
       expect(records).toHaveLength(2)
       expect(featureCollection.features).toHaveLength(1)
     })
