+++ src/types.ts
@@ -32,3 +32,12 @@ export interface AirtableGeoJSONErrors<F> {
   missingGeocodes: Airtable.Record<F>[]
   invalidGeocodes: Airtable.Record<F>[]
 }
+
+/**
+ * Used to assert that our generic types are key-value maps
+ * that contain at least one string -> string mapping, namely
+ * the one for the geocoded field name -> geocoder cache value
+ */
+export interface StringMap {
+  [key: string]: string
+}
